Update callbacks documentation; Remove CallbackFactory as it was superfluous. Move Callback classes to listener interfaces
